Responsive design for cache page.
